+++ packages/visual-grid-client/src/sdk/wrapperUtils.js
@@ -92,14 +92,7 @@ function configureWrappers({
   }
 }
 
-function openWrappers({
-  wrappers,
-  browsers,
-  appName,
-  testName,
-  eyesTransactionThroat,
-  skipStartingSession,
-}) {
+function openWrappers({wrappers, browsers, appName, testName, eyesTransactionThroat}) {
   const openPromisesAndResolves = wrappers.map(() => {
     let resolve
     const promise = new Promise(r => (resolve = r))
@@ -109,9 +102,7 @@ function openWrappers({
     .map((wrapper, i) => {
       const viewportSize = browsers[i].width && new RectangleSize(browsers[i])
       return eyesTransactionThroat(() =>
-        wrapper
-          .open({appName, testName, viewportSize, skipStartingSession})
-          .then(openPromisesAndResolves[i].resolve),
+        wrapper.open({appName, testName, viewportSize}).then(openPromisesAndResolves[i].resolve),
       )
     })
     .reduce(